Conflict with Lombok resolved
