+++ src/main/java/ru/skypro/homework/config/WebSecurityConfig.java
@@ -9,7 +9,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
-import ru.skypro.homework.dto.register.Role;
+import ru.skypro.homework.dto.auth.Role;
 
 import static org.springframework.security.config.Customizer.withDefaults;
