+++ src/main/java/ru/skypro/homework/dto/ads/ExtendedAdDto.java
@@ -4,7 +4,6 @@ import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
 @Data
-@RequiredArgsConstructor
 public class ExtendedAdDto {
     private Integer pk;
     private String authorFirstName;
