+++ src/main/java/ru/skypro/homework/dto/auth/RegisterDto.java
@@ -1,4 +1,4 @@
-package ru.skypro.homework.dto.users;
+package ru.skypro.homework.dto.auth;
 
 import lombok.Data;
 
@@ -6,12 +6,17 @@ import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
 @Data
-public class UpdateUserDto {
-    @Size(min = 3, max = 10)
+public class RegisterDto {
+    @Size(min = 4, max = 32)
+    private String username;
+    @Size(min = 8, max = 16)
+    private String password;
+    @Size(min = 2, max = 16)
     private String firstName;
-    @Size(min = 3, max = 10)
+    @Size(min = 2, max = 16)
     private String lastName;
     @Pattern(regexp = "\\+7\\s?\\(?\\d{3}\\)?\\s?\\d{3}-?\\d{2}-?\\d{2}", message = "Неверный формат введенного номера телефона")
     private String phone;
+    private Role role;
 
 }
