+++ src/main/java/ru/skypro/homework/dto/auth/Role.java
@@ -1,0 +1,5 @@
+package ru.skypro.homework.dto.auth;
+
+public enum Role {
+    USER, ADMIN
+}
