+++ src/main/java/ru/skypro/homework/service/auth/AuthService.java
@@ -1,6 +1,6 @@
 package ru.skypro.homework.service.auth;
 
-import ru.skypro.homework.dto.register.RegisterDto;
+import ru.skypro.homework.dto.auth.RegisterDto;
 
 public interface AuthService {
     boolean login(String userName, String password);
