+++ src/main/java/ru/skypro/homework/service/auth/impl/AuthServiceImpl.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.stereotype.Service;
-import ru.skypro.homework.dto.register.RegisterDto;
+import ru.skypro.homework.dto.auth.RegisterDto;
 import ru.skypro.homework.service.auth.AuthService;
 
 @Service
